requirements.txt
Django==1.11.14
# django-classy-tags==0.7.2
# django-cms==3.5.2
https://github.com/divio/django-cms/archive/release/3.5.x.zip
django-sekizai==0.10.0
# django-treebeard==4.0
djangocms-admin-style==1.2.6
djangocms-text-ckeditor==3.6.0
ipdb
uWSGI==2.0.15
dj-database-url==0.4.1
psycopg2==2.6.1
django-filer==1.3.2
cmsplugin-filer==1.1.3
# django-reversion<=1.11
djangocms-page-meta==0.8.0
django-widget-tweaks
geopy
django-admin-sortable==2.1.7